Added some new window
I added the window about the "Instruction" and "About" menu items.
Whenever one of this is triggered, a new window is opened giving some
information. When "Instructions" is triggered, some instructions about
the general use is given. Instead, when "About" is triggered,  some
information about the scope of the project and the main reasons for
developing it are displayed!